drivers: pwm_qmsi: add type cast to prevent overflow
Add type cast to prevent overflow.

In the pwm driver function pwm_qmsi_get_cycles_per_sec(),
there is a multiplication which takes ticks_per_us (32 bit
unsigned integer) and USEC_PER_SEC. Practically, overflow
will not happen, since the sys clock is not that high. But,
it can trigger an overflow alarm by automatic tools.

+++ drivers/pwm/pwm_qmsi.c
@@ -401,7 +401,7 @@ static int pwm_qmsi_get_cycles_per_sec(struct device *dev, uint32_t pwm,
 		return -EINVAL;
 	}
 
-	*cycles = clk_sys_get_ticks_per_us() * USEC_PER_SEC;
+	*cycles = (uint64_t)clk_sys_get_ticks_per_us() * USEC_PER_SEC;
 
 	return 0;
 }
